+++ src/optimagic/optimization/optimize.py
@@ -12,21 +12,14 @@ is then passed to `_optimize` which handles the optimization logic.
 
 """
 
-import warnings
-from pathlib import Path
-from typing import cast
+from typing import Any, cast
 
 from optimagic.batch_evaluators import process_batch_evaluator
 from optimagic.exceptions import (
     InvalidFunctionError,
 )
-from optimagic.logging.create_tables import (
-    make_optimization_iteration_table,
-    make_optimization_problem_table,
-    make_steps_table,
-)
-from optimagic.logging.load_database import load_database
-from optimagic.logging.write_to_database import append_row
+from optimagic.logging.logger import LogReader, LogStore
+from optimagic.logging.types import ProblemInitialization
 from optimagic.optimization.create_optimization_problem import (
     OptimizationProblem,
     create_optimization_problem,
@@ -70,8 +63,7 @@ def maximize(
     fun_and_jac=None,
     fun_and_jac_kwargs=None,
     numdiff_options=None,
-    logging=False,
-    log_options=None,
+    logging=None,
     error_handling="raise",
     error_penalty=None,
     scaling=False,
@@ -96,6 +88,7 @@ def maximize(
     derivative_kwargs=None,
     criterion_and_derivative=None,
     criterion_and_derivative_kwargs=None,
+    log_options=None,
     lower_bounds=None,
     upper_bounds=None,
     soft_lower_bounds=None,
@@ -185,8 +178,7 @@ def minimize(
     fun_and_jac=None,
     fun_and_jac_kwargs=None,
     numdiff_options=None,
-    logging=False,
-    log_options=None,
+    logging=None,
     error_handling="raise",
     error_penalty=None,
     scaling=False,
@@ -212,6 +204,7 @@ def minimize(
     criterion_and_derivative=None,
     criterion_and_derivative_kwargs=None,
     lower_bounds=None,
+    log_options=None,
     upper_bounds=None,
     soft_lower_bounds=None,
     soft_upper_bounds=None,
@@ -251,7 +244,6 @@ def minimize(
         fun_and_jac_kwargs=fun_and_jac_kwargs,
         numdiff_options=numdiff_options,
         logging=logging,
-        log_options=log_options,
         error_handling=error_handling,
         error_penalty=error_penalty,
         scaling=scaling,
@@ -277,6 +269,7 @@ def minimize(
         criterion_and_derivative=criterion_and_derivative,
         criterion_and_derivative_kwargs=criterion_and_derivative_kwargs,
         lower_bounds=lower_bounds,
+        log_options=log_options,
         upper_bounds=upper_bounds,
         soft_lower_bounds=soft_lower_bounds,
         soft_upper_bounds=soft_upper_bounds,
@@ -292,8 +285,6 @@ def _optimize(problem: OptimizationProblem) -> OptimizeResult:
     # Split constraints into nonlinear and reparametrization parts
     # ==================================================================================
     constraints = problem.constraints
-    if isinstance(constraints, dict):
-        constraints = [constraints]
 
     nonlinear_constraints = [c for c in constraints if c["type"] == "nonlinear"]
 
@@ -355,21 +346,14 @@ def _optimize(problem: OptimizationProblem) -> OptimizeResult:
     # ==================================================================================
     # initialize the log database
     # ==================================================================================
-    if problem.logging:
-        # TODO: We want to remove the optimization_problem table completely but we
-        # probably do need to store the start parameters in the database because it is
-        # used by the log reader.
-        problem_data = {
-            "direction": problem.direction.value,
-            "params": problem.params,
-        }
-        database = _create_and_initialize_database(
-            logging=problem.logging,
-            log_options=problem.log_options,
-            problem_data=problem_data,
-        )
+    logger: LogStore[Any, Any] | None
+
+    if problem.logger:
+        logger = LogStore.from_options(problem.logger)
+        problem_data = ProblemInitialization(problem.direction, problem.params)
+        logger.problem_store.insert(problem_data)
     else:
-        database = None
+        logger = None
 
     # ==================================================================================
     # Do some things that require internal parameters or bounds
@@ -443,8 +427,7 @@ def _optimize(problem: OptimizationProblem) -> OptimizeResult:
         # TODO: Actually use the step ids
         step_ids = log_scheduled_steps_and_get_ids(  # noqa: F841
             steps=steps,
-            logging=problem.logging,
-            database=database,
+            logging=logger,
         )
 
         raw_res = problem.algorithm.solve_internal_problem(internal_problem, x)
@@ -467,8 +450,7 @@ def _optimize(problem: OptimizationProblem) -> OptimizeResult:
             x=x,
             sampling_bounds=sampling_bounds,
             options=multistart_options,
-            logging=problem.logging,
-            database=database,
+            logging=logger,
             error_handling=problem.error_handling,
         )
 
@@ -479,6 +461,7 @@ def _optimize(problem: OptimizationProblem) -> OptimizeResult:
     _scalar_start_criterion = cast(
         float, first_crit_eval.internal_value(AggregationLevel.SCALAR)
     )
+    log_reader: LogReader[Any] | None
 
     extra_fields = ExtraResultFields(
         start_fun=_scalar_start_criterion,
@@ -503,55 +486,12 @@ def _optimize(problem: OptimizationProblem) -> OptimizeResult:
             extra_fields=extra_fields,
         )
 
+    if logger is not None:
+        assert problem.logger is not None
+        log_reader = LogReader.from_options(problem.logger)
+    else:
+        log_reader = None
+
+    res.logger = log_reader
+
     return res
-
-
-def _create_and_initialize_database(logging, log_options, problem_data):
-    """Create and initialize to sqlite database for logging."""
-    path = Path(logging)
-    fast_logging = log_options.get("fast_logging", False)
-    if_table_exists = log_options.get("if_table_exists", "extend")
-    if_database_exists = log_options.get("if_database_exists", "extend")
-
-    if "if_exists" in log_options and "if_table_exists" not in log_options:
-        warnings.warn("The log_option 'if_exists' was renamed to 'if_table_exists'.")
-
-    if logging.exists():
-        if if_database_exists == "raise":
-            raise FileExistsError(
-                f"The database {logging} already exists and the log_option "
-                "'if_database_exists' is set to 'raise'"
-            )
-        elif if_database_exists == "replace":
-            logging.unlink()
-
-    database = load_database(path_or_database=path, fast_logging=fast_logging)
-
-    # create the optimization_iterations table
-    make_optimization_iteration_table(
-        database=database,
-        if_exists=if_table_exists,
-    )
-
-    # create and initialize the steps table; This is alway extended if it exists.
-    make_steps_table(database, if_exists=if_table_exists)
-
-    # create_and_initialize the optimization_problem table
-    make_optimization_problem_table(database, if_exists=if_table_exists)
-
-    not_saved = [
-        "criterion",
-        "criterion_kwargs",
-        "constraints",
-        "derivative",
-        "derivative_kwargs",
-        "criterion_and_derivative",
-        "criterion_and_derivative_kwargs",
-    ]
-    problem_data = {
-        key: val for key, val in problem_data.items() if key not in not_saved
-    }
-
-    append_row(problem_data, "optimization_problem", database=database)
-
-    return database